reorganize song components
