+++ src/components/SongList.js
@@ -1,38 +1,36 @@
 import React from "react";
-import Song from "./Song";
 import useApp from "./useApp";
+import Song from "./Song";
 
 
 const SongList = () => {
+  const { songs, filteredGenre, filteredRating } = useApp();
 
-  const { sortList, sortRating } = useApp();
+  let filteredList = [...songs];
+
+  // filtering for genre
+  if (filteredGenre !== "All") {
+    filteredList = filteredList.filter((song) => {
+      return song.genre === filteredGenre;
+    });
+  }
+
+  // filtering for rating
+  if (filteredRating !== "All") {
+    filteredList = filteredList.filter((song) => {
+      return song.rating === parseInt(filteredRating);
+    });
+  }
+
+  const songsList = filteredList.map((song) => (
+    <Song key={song.id} song={song} />
+  ));
 
   return (
-    <div className="song-list">
-      <div className="song-list__message">* to sort click on column names</div>
-      <table>
-        <thead>
-          <tr>
-            <th>
-              <span onClick={sortList}>Title</span>
-            </th>
-            <th>
-              <span onClick={sortList}>Artist</span>
-            </th>
-            <th>
-              <span onClick={sortList}>Genre</span>
-            </th>
-            <th>
-              <span onClick={sortRating}>Rating</span>
-            </th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          <Song />
-        </tbody>
-      </table>
-    </div>
+    <>
+      {songsList}
+    </>
+    
   );
 };
